Initial commit for spring-boot-app-v1

+++ spring-boot-basic-training-v1/mypro/src/main/java/com/example/mypro/service/StudentService.java
@@ -2,8 +2,11 @@ package com.example.mypro.service;
 import com.example.mypro.model.Student;
 import com.example.mypro.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
+@Service
 public class StudentService {
     @Autowired
     private StudentRepository studentRepository; // Assuming you've created a StudentRepository interface
@@ -20,10 +23,12 @@ public class StudentService {
         return studentRepository.save(student);
     }
 
-    public void deleteStudent(Long id) {
-        studentRepository.deleteById(id);
-    }
-
-    public void deleteStudentById(Long id) {
+    public boolean deleteStudentById(Long id) {
+        if (studentRepository.existsById(id)) {
+            studentRepository.deleteById(id)
+            ;
+            return true;
+        }
+        return false;
     }
 }
